RQ-216 Renamed folder | also updated starter URL
