+++ app/src/views/features/api-client/client-view/response/responseTabs.scss
