+++ app/src/views/features/api-client/sidebar/apiClientSidebar.scss
@@ -27,6 +27,7 @@
     width: 100%;
     height: 100%;
     flex: auto;
+    overflow: auto;
 
     &.ant-menu {
       background: transparent;
